requirements.txt
discord
pandas
random
numpy
asyncio
pathlib
python-dotenv